requirements.txt
Django >= 1.5, < 1.6
django-annoying >= 0.7.6
django-debug-toolbar >= 0.9.4
south >= 0.7.6
pillow >= 2.0.0
factory_boy >= 2.0.0
